Revert to default kernel for now
Need to figure out how to get akmods, kernel headers, w/e to work with
the audinux kernels, but I feel like it's been impeding my progress too
much for something that's debatably not necessary?

Still would like to implement a high-performance kernel but that's being
placed on the backburner until this project is a thing that actually
exists and is able to be tested in any capacity.